Collective updates

+++ R/clustering.R
@@ -25,7 +25,9 @@
 #' @param nRandomStarts Integer number of random starts. Will pick the
 #' membership with highest quality to return. Default \code{10}.
 #' @param nIterations Integer, maximal number of iterations per random start.
-#' Default \code{100}.
+#' Default \code{5}.
+#' @param method Community detection algorithm to use. Choose from
+#' \code{"leiden"} or \code{"louvain"}. Default \code{"leiden"}.
 #' @param useDims Indices of factors to use for clustering. Default \code{NULL}
 #' uses all available factors.
 #' @param groupSingletons Whether to group single cells that make up their own
@@ -37,127 +39,37 @@
 #' @param seed Seed of the random number generator. Default \code{1}.
 #' @param verbose Logical. Whether to show information of the progress. Default
 #' \code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.
-#' @param partitionType Choose from \code{"ModularityVertexPartition",
-#' "RBConfigurationVertexPartition", "RBERVertexPartition",
-#' "SignificanceVertexPartition", "CPMVertexPartition",
-#' "SurpriseVertexPartition"}. See
-#' \code{\link[leidenbase]{leiden_find_partition}} for detail. Default
-#' \code{"ModularityVertexPartition"}.
-#' @param ... Additional arguments passed to
-#' \code{\link[leidenbase]{leiden_find_partition}}, including
-#' \code{initial_membership} and \code{node_sizes}.
 #' @return \code{object} with refined cluster assignment updated in
 #' \code{clusterName} variable in \code{cellMeta} slot. Can be fetched
 #' with \code{object[[clusterName]]}
-#' @rdname clustering
+#' @rdname runCluster
 #' @export
 #' @examples
-#' pbmcPlot <- runLeidenCluster(pbmcPlot)
+#' pbmcPlot <- runCluster(pbmcPlot)
 #' head(pbmcPlot$leiden_cluster)
-#' pbmcPlot <- runLouvainCluster(pbmcPlot)
+#' pbmcPlot <- runCluster(pbmcPlot, method = "louvain")
 #' head(pbmcPlot$louvain_cluster)
-runLeidenCluster <- function(
-        object,
-        resolution = 1.0,
-        nNeighbors = 20,
-        partitionType = c("ModularityVertexPartition",
-                          "RBConfigurationVertexPartition",
-                          "RBERVertexPartition",
-                          "SignificanceVertexPartition",
-                          "CPMVertexPartition",
-                          "SurpriseVertexPartition"),
-        prune = 1 / 15,
-        eps = 0.1,
-        nRandomStarts = 10,
-        nIterations = 100,
-        useDims = NULL,
-        groupSingletons = TRUE,
-        clusterName = "leiden_cluster",
-        seed = 1,
-        verbose = getOption("ligerVerbose"),
-        ...
-) {
-    partitionType <- match.arg(partitionType)
-    object <- recordCommand(object,
-                            dependencies = c("RANN", "leidenbase", "igraph"))
-    H.norm <- getMatrix(object, "H.norm")
-    if (is.null(H.norm)) {
-        type <- " unnormalized "
-        H.norm <- Reduce(cbind, getMatrix(object, "H"))
-    } else type <- " quantile normalized "
-    if (is.null(H.norm))
-        stop("No factor loading ('H.norm' or 'H') found in `object`.")
-    if (type == " unnormalized ") H.norm <- t(H.norm)
-
-    if (!is.null(useDims)) H.norm <- H.norm[, useDims]
-
-    if (isTRUE(verbose))
-        .log("Leiden clustering on", type, "cell factor loadings...")
-    knn <- RANN::nn2(H.norm, k = nNeighbors, eps = eps)
-    snn <- ComputeSNN(knn$nn.idx, prune = prune)
-    snnSummary <- summary(snn)
-    edges <- as.vector(t(snnSummary[,c(1, 2)]))
-    g <- igraph::graph(edges = edges, n = nrow(snn),directed = FALSE)
-    igraph::E(g)$weight <- snnSummary[, 3]
-
-    set.seed(seed)
-    maxQuality <- -1
-    if (isTRUE(verbose))
-        pb <- utils::txtProgressBar(0, nRandomStarts, style = 3)
-    for (i in seq(nRandomStarts)) {
-        seed <- sample(1000, 1)
-        part <- leidenbase::leiden_find_partition(
-            igraph = g,
-            partition_type = partitionType,
-            resolution_parameter = resolution,
-            edge_weights = igraph::E(g)$weight,
-            num_iter = nIterations,
-            verbose = FALSE,
-            seed = seed,
-            ...
-        )
-        #if (is.null(part$quality)) {
-        #    clusts <- part$membership
-        #    if (isTRUE(verbose)) utils::setTxtProgressBar(pb, nRandomStarts)
-        #    break
-        #}
-        if (part$quality > maxQuality) {
-            clusts <- part$membership
-            maxQuality <- part$quality
-        }
-        if (isTRUE(verbose)) utils::setTxtProgressBar(pb, i)
-    }
-    if (isTRUE(verbose)) cat("\n")
-    names(clusts) <- colnames(object)
-    rownames(snn) <- colnames(object)
-    colnames(snn) <- colnames(object)
-
-    clusts <- groupSingletons(ids = clusts, SNN = snn,
-                              groupSingletons = groupSingletons,
-                              verbose = verbose)
-    if (is.numeric(clusts)) clusts <- clusts - 1
-    clusts <- factor(clusts)
-    cellMeta(object, clusterName, check = FALSE) <- clusts
-    return(object)
-}
-
-#' @rdname clustering
-#' @export
-runLouvainCluster <- function(
+runCluster <- function(
         object,
         resolution = 1.0,
         nNeighbors = 20,
         prune = 1 / 15,
         eps = 0.1,
         nRandomStarts = 10,
-        nIterations = 100,
+        nIterations = 5,
+        method = c("leiden", "louvain"),
         useDims = NULL,
         groupSingletons = TRUE,
-        clusterName = "louvain_cluster",
+        clusterName = paste0(method, "_cluster"),
         seed = 1,
         verbose = getOption("ligerVerbose")
 ) {
-    object <- recordCommand(object, dependencies = "RANN")
+    method <- match.arg(method)
+    object <- switch(method,
+        leiden = recordCommand(object, dependencies = c("RANN", "leidenAlg")),
+        louvain = recordCommand(object, dependencies = c("RANN"))
+    )
+
     H.norm <- getMatrix(object, "H.norm")
     if (is.null(H.norm)) {
         type <- " unnormalized "
@@ -165,40 +77,53 @@ runLouvainCluster <- function(
     } else type <- " quantile normalized "
     if (is.null(H.norm))
         stop("No factor loading ('H.norm' or 'H') found in `object`.")
-    # Not transposing when cbind'ing becausing `t(NULL)` causes error
     if (type == " unnormalized ") H.norm <- t(H.norm)
-
-    edgeOutPath <- paste0("edge_", sub("\\s", "_", Sys.time()), '.txt')
-    edgeOutPath <- gsub("-", "", edgeOutPath)
-    edgeOutPath <- gsub(":", "", edgeOutPath)
 
     if (!is.null(useDims)) H.norm <- H.norm[, useDims]
 
     if (isTRUE(verbose))
-        .log("Louvain clustering on", type, "cell factor loadings...")
+        .log(method, " clustering on", type, "cell factor loadings...")
     knn <- RANN::nn2(H.norm, k = nNeighbors, eps = eps)
     snn <- ComputeSNN(knn$nn.idx, prune = prune)
-    WriteEdgeFile(snn, edgeOutPath, display_progress = FALSE)
-    clusts <- RunModularityClusteringCpp(
-        snn,
-        modularityFunction = 1,
-        resolution = resolution,
-        nRandomStarts = nRandomStarts,
-        nIterations = nIterations,
-        algorithm = 1,
-        randomSeed = seed,
-        printOutput = FALSE,
-        edgefilename = edgeOutPath
-    )
+    if (!is.null(seed)) set.seed(seed)
+    if (method == "leiden") {
+        snnSummary <- summary(snn)
+        edgelist <- as.vector(t(snnSummary[, c(2, 1)])) - 1
+        edgelist_length <- length(edgelist)
+        edge_weights <- snnSummary[,3]
+        clusts <- leidenAlg::find_partition_with_rep_rcpp(
+            edgelist = edgelist, edgelist_length = edgelist_length,
+            num_vertices = nrow(H.norm), direction = FALSE,
+            edge_weights = edge_weights, resolution = resolution,
+            niter = nIterations, nrep = nRandomStarts
+        )
+    } else {
+        edgeOutPath <- paste0("edge_", sub("\\s", "_", Sys.time()), '.txt')
+        edgeOutPath <- gsub("-", "", edgeOutPath)
+        edgeOutPath <- gsub(":", "", edgeOutPath)
+        WriteEdgeFile(snn, edgeOutPath, display_progress = FALSE)
+        clusts <- RunModularityClusteringCpp(
+            snn,
+            modularityFunction = 1,
+            resolution = resolution,
+            nRandomStarts = nRandomStarts,
+            nIterations = nIterations,
+            algorithm = 1,
+            randomSeed = seed,
+            printOutput = TRUE,
+            edgefilename = edgeOutPath
+        )
+        unlink(edgeOutPath)
+    }
+    clusts <- .labelClustBySize(clusts)
     names(clusts) <- colnames(object)
     rownames(snn) <- colnames(object)
     colnames(snn) <- colnames(object)
-    # clusts must not be a factor at this point
     clusts <- groupSingletons(ids = clusts, SNN = snn,
                               groupSingletons = groupSingletons,
                               verbose = verbose)
-    cellMeta(object, columns = clusterName, check = FALSE) <- factor(clusts)
-    unlink(edgeOutPath)
+    cellMeta(object, clusterName, check = FALSE) <- clusts
+    object@uns$defaultCluster <- object@uns$defaultCluster %||% clusterName
     return(object)
 }
 
@@ -234,10 +159,11 @@ NULL
 
 #' @rdname rliger2-deprecated
 #' @section \code{louvainCluster}:
-#' For \code{louvainCluster}, use \code{\link{runLouvainCluster}} as the
-#' replacement, while \code{\link{runLeidenCluster}} is more recommended.
+#' For \code{louvainCluster}, use \code{\link{runCluster}(method = "louvain")}
+#' as the replacement, while \code{\link{runCluster}} with default
+#' \code{method = "leiden"} is more recommended.
 #' @export
-louvainCluster <- function(
+louvainCluster <- function( # nocov start
         object,
         resolution = 1.0,
         k = 20,
@@ -249,15 +175,16 @@ louvainCluster <- function(
         verbose = getOption("ligerVerbose"),
         dims.use = NULL
 ) {
-    lifecycle::deprecate_warn("1.99.0", "louvainCluster()",
-                              "runLouvainCluster()")
-    runLouvainCluster(
-        object, resolution = resolution, nNeighbors = k, prune = prune,
-        eps = eps, nRandomStarts = nRandomStarts, nIterations = nIterations,
-        useDims = dims.use, groupSingletons = TRUE,
+    lifecycle::deprecate_warn(
+        "1.99.0", "louvainCluster()",
+        details = "Please use `runCluster()` with `method = 'louvain'` instead.")
+    runCluster(
+        object, method = "louvain", resolution = resolution, nNeighbors = k,
+        prune = prune, eps = eps, nRandomStarts = nRandomStarts,
+        nIterations = nIterations, useDims = dims.use, groupSingletons = TRUE,
         clusterName = "louvain_cluster", seed = random.seed, verbose = verbose
     )
-}
+} # nocov end
 
 # Group single cells that make up their own cluster in with the cluster they are
 # most connected to. (Adopted from Seurat v3)
@@ -276,23 +203,32 @@ groupSingletons <- function(
         verbose = FALSE
 ) {
     # identify singletons
+    ids <- droplevels(ids)
     singletons <- names(which(table(ids) == 1))
-    singletons <- intersect(unique(ids), singletons)
-    if (!isTRUE(groupSingletons)) {
-        ids[ids %in% singletons] <- "singleton"
+    singletons <- intersect(levels(ids), singletons)
+    if (length(singletons) == 0) {
         return(ids)
+    }
+    if (!isTRUE(groupSingletons)) {
+        if (isTRUE(verbose)) {
+            .log(length(singletons), " singletons identified. ")
+        }
+        ids <- as.character(ids)
+        ids[ids %in% singletons] <- "singleton"
+        return(factor(ids))
     }
     # calculate connectivity of singletons to other clusters, add singleton
     # to cluster it is most connected to
-    clusterNames <- unique(ids)
+    clusterNames <- levels(ids)
     clusterNames <- setdiff(clusterNames, singletons)
     connectivity <- vector(mode = "numeric", length = length(clusterNames))
     names(connectivity) <- clusterNames
+    ids <- as.character(ids)
     for (i in singletons) {
-        i.cells <- names(which(ids == i))
+        i.cells <- which(ids == i)
         for (j in clusterNames) {
-            j.cells <- names(which(ids == j))
-            subSNN <- SNN[i.cells, j.cells]
+            j.cells <- which(ids == j)
+            subSNN <- SNN[i.cells, j.cells, drop = FALSE]
             # to match previous behavior, random seed being set in WhichCells
             set.seed(1)
             if (is.object(subSNN))
@@ -305,8 +241,69 @@ groupSingletons <- function(
         closest_cluster <- sample(names(connectivity[mi]), 1)
         ids[i.cells] <- closest_cluster
     }
-    if (length(singletons) > 0 && isTRUE(verbose))
+    ids <- factor(ids)
+    if (isTRUE(verbose))
         .log(length(singletons), " singletons identified. ",
-             length(unique(ids)), " final clusters.")
+             length(levels(ids)), " final clusters.")
     return(ids)
+}
+
+.labelClustBySize <- function(clusts) {
+    clusts <- as.character(clusts)
+    count <- data.frame(table(clusts))
+    count <- count[order(count$Freq, decreasing = TRUE),]
+    map <- seq(0, nrow(count) - 1)
+    names(map) <- count[[1]]
+    newClusts <- map[clusts]
+    newClusts <- factor(newClusts)
+    return(newClusts)
+}
+
+
+
+#' Create new variable from categories in cellMeta
+#' @description
+#' Designed for fast variable creation when a new variable is going to be
+#' created from existing variable. For example, multiple samples can be mapped
+#' to the same study design condition, clusters can be mapped to cell types.
+#' @param object A \linkS4class{liger} object.
+#' @param from The name of the original variable to be mapped from.
+#' @param newTo The name of the new variable to store the mapped result. Default
+#' \code{NULL} returns the new variable (factor class).
+#' @param ... Mapping criteria, argument names are original existing categories
+#' in the \code{from} and values are new categories in the new variable.
+#' @return When \code{newTo = NULL}, a factor object of the new variable.
+#' Otherwise, the input object with variable \code{newTo} updated in
+#' \code{cellMeta(object)}.
+#' @export
+#' @examples
+#' pbmc <- mapCellMeta(pbmc, from = "dataset", newTo = "modal",
+#'                     ctrl = "rna", stim = "rna")
+mapCellMeta <- function(
+        object,
+        from,
+        newTo = NULL,
+        ...
+) {
+    object <- recordCommand(object, ...)
+    from <- cellMeta(object, from)
+    if (!is.factor(from)) stop("`from` must be a factor class variable.")
+    mapping <- list(...)
+    fromCats <- names(mapping)
+    notFound <- fromCats[!fromCats %in% levels(from)]
+    if (length(notFound) > 0) {
+        stop("The following categories requested not found: ",
+             paste0(notFound, collapse = ", "))
+    }
+
+    toCats <- unlist(mapping)
+    unchangedCats <- levels(from)
+    unchangedCats <- unchangedCats[!unchangedCats %in% fromCats]
+    names(unchangedCats) <- unchangedCats
+    if (length(unchangedCats) > 0) toCats <- c(toCats, unchangedCats)
+    to <- toCats[as.character(from)]
+    to <- factor(unname(to), levels = unique(toCats))
+    if (is.null(newTo)) return(to)
+    cellMeta(object, newTo) <- to
+    return(object)
 }
